requirement.txt
anytree==2.6.0
contourpy==1.3.1
cycler==0.12.1
fonttools==4.56.0
graphviz==0.10.1
kiwisolver==1.4.8
matplotlib==3.10.0
mysql-connector-python==9.2.0
numpy==2.2.3
packaging==24.2
pillow==11.1.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
setuptools==40.8.0
six==1.17.0
treeview==1.2
ttkthemes==3.2.2
